requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
priority==2.0.0
python-dotenv==0.21.0
python-mpd2==3.0.5
pytube==12.1.0
quart==0.18.3
toml==0.10.2
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
